% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine a transmission tree and a phylogeny into a colored phylogeny}
\usage{
combine(ttree, ptree)
}
\arguments{
\item{ttree}{Transmission tree}

\item{ptree}{Phylogenetic tree}
}
\value{
colored tree
}
\description{
Combine a transmission tree and a phylogeny into a colored phylogeny
}
