% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{o2mod.transphylo}
\alias{o2mod.transphylo}
\title{Run outbreaker2 using the TransPhylo model}
\usage{
o2mod.transphylo(data, config = NULL, priors = NULL, likelihoods = NULL,
  moves = NULL)
}
\arguments{
\item{data}{a list of named items containing input data as returned by
\code{\link[outbreaker2]{outbreaker_data}}, with an additional item
\code{ptree} providing the phylogenetic tree}

\item{config}{a set of settings as returned by
\code{\link[outbreaker2]{create_config}}}

\item{priors}{a set of log-prior functions as returned by
\code{\link[outbreaker2]{custom_priors}}}

\item{likelihoods}{a set of log-likelihood functions as returned by
\code{\link[outbreaker2]{custom_likelihoods}}}

\item{moves}{a set of movement functions as returned by
\code{\link[outbreaker2]{custom_moves}}}
}
\value{
an \code{outbreaker_chains} object
}
\description{
Run outbreaker2 using the TransPhylo model
}
