% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{o2mod.transphylo}
\alias{o2mod.transphylo}
\title{Run outbreaker2 using the TransPhylo model}
\usage{
o2mod.transphylo(data, n_iter = 1000, sample_every = 1)
}
\arguments{
\item{data}{Data}

\item{n_iter}{Number of iterations}

\item{sample_every}{Thining interval}
}
\value{
results of the run as an outbreaker2 output object
}
\description{
Run outbreaker2 using the TransPhylo model
}

